functions/requirements.txt
firebase_functions~=0.1.0
flask
gunicorn
google-api-python-client==2.1.0
black
requests
google-cloud-firestore
cryptography
app-store-server-library
python-dotenv